+++ backend/main.py
@@ -1,12 +1,11 @@
 from contextlib import asynccontextmanager
 
-from fastapi import FastAPI, HTTPException
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from pydantic import BaseModel, EmailStr
-from sqlalchemy.exc import IntegrityError
 
 import models  # ensures all models are registered with Base before create_all
-from database import AsyncSessionLocal, engine
+from database import engine
+from routes import health_router, waitlist_router
 
 
 @asynccontextmanager
@@ -25,28 +24,5 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-
-@app.get("/")
-async def root():
-    return {"message": "Veriva API is running"}
-
-
-@app.get("/health")
-async def health():
-    return {"status": "ok"}
-
-
-class WaitlistRequest(BaseModel):
-    email: EmailStr
-
-
-@app.post("/waitlist")
-async def join_waitlist(body: WaitlistRequest):
-    async with AsyncSessionLocal() as session:
-        waitlist_entry = models.WaitList(email=body.email)
-        session.add(waitlist_entry)
-        try:
-            await session.commit()
-        except IntegrityError:
-            raise HTTPException(status_code=409, detail="Email already on waitlist")
-    return {"message": "Email added to waitlist"}
+app.include_router(health_router)
+app.include_router(waitlist_router)
