+++ backend/routes/__init__.py
@@ -1,0 +1,4 @@
+from .health import router as health_router
+from .waitlist import router as waitlist_router
+
+__all__ = ["health_router", "waitlist_router"]
